+++ angular.json
@@ -96,7 +96,8 @@
               "src/assets"
             ],
             "styles": [
-              "node_modules/bootstrap/dist/css/bootstrap-theme.min.css",
+              "../node_modules/bootstrap/dist/css/bootstrap-theme.min.css",
+              "node_modules/angular-material/angular-material.min.css",
               "src/styles.scss"
             ],
             "scripts": []
