+++ db.json
@@ -9,11 +9,6 @@
       "id": 3,
       "title": "andreas",
       "price": "maik"
-    },
-    {
-      "id": 14,
-      "title": "test",
-      "price": "test"
     }
   ],
   "comments": [
