+++ src/app/app.module.ts
@@ -1,27 +1,34 @@
 import { NgModule } from '@angular/core';
+import { AppComponent } from './app.component';
 
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { MatButtonModule } from '@angular/material/button';
 import { MatIconModule } from '@angular/material/icon';
-
 import { BrowserModule } from '@angular/platform-browser';
-import { ReactiveFormsModule }  from '@angular/forms';
+import { ReactiveFormsModule } from '@angular/forms';
 import { MatCardModule } from '@angular/material/card';
 import { HttpClientModule } from '@angular/common/http';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
-
-import { AppComponent } from './app.component';
-import { TopBarComponent } from './top-bar/top-bar.component';
-import { ProductListComponent } from './product-list/product-list.component';
+import { MatTabsModule } from '@angular/material/tabs';
 import { FormsModule } from '@angular/forms';
-import { OrderListComponent } from './order-list/order-list.component';
+import { Routes } from '@angular/router';
+
 import { HeaderComponent } from './header/header.component';
 import { OrdersComponent } from './orders/orders.component';
-import { OrderDetailComponent } from './orders/order-detail/order-detail.component';
 import { OrderItemComponent } from './orders/order-list/order-item/order-item.component';
 import { ProductsComponent } from './products/products.component';
-import { ProductDetailComponent } from './products/product-detail/product-detail.component';
 import { ProductItemComponent } from './products/product-list/product-item/product-item.component';
+import { OrderListComponent } from './orders/order-list/order-list.component';
+import { ProductListComponent } from './products/product-list/product-list.component';
+import { HomeComponent } from './home/home.component';
+
+const routes: Routes = [
+  { path: '', redirectTo: '/home', pathMatch: 'full' },
+  { path: 'home', component: HomeComponent },
+  { path: 'product', component: ProductsComponent },
+  { path: 'order', component: OrdersComponent },
+];
 
 @NgModule({
   imports: [
@@ -33,20 +40,20 @@ import { ProductItemComponent } from './products/product-list/product-item/produ
     MatFormFieldModule,
     MatInputModule,
     HttpClientModule,
-    ReactiveFormsModule
+    ReactiveFormsModule,
+    MatTabsModule,
+    BrowserAnimationsModule
   ],
   declarations: [
     AppComponent,
-    TopBarComponent,
-    ProductListComponent,
-    OrderListComponent,
     HeaderComponent,
     OrdersComponent,
-    OrderDetailComponent,
     OrderItemComponent,
     ProductsComponent,
-    ProductDetailComponent,
     ProductItemComponent,
+    OrderListComponent,
+    ProductListComponent,
+    HomeComponent
   ],
   bootstrap: [
     AppComponent
