+++ src/app/header/header.component.html
@@ -1,1 +1,12 @@
-<h3>header works!</h3>
+
+<!-- <button mat-raised-button
+        class="example-action-button"
+        (click)="toggleBackground()">
+  Toggle background
+</button>
+
+<nav mat-tab-nav-bar>
+  <a mat-tab-link *ngFor="let link of links"
+     (click)="activeLink = link"
+     [active]="activeLink == link"> {{link}} </a>
+</nav> -->
