+++ src/app/home/home.component.html
@@ -1,0 +1,1 @@
+<h1>"Small steps, big coder."</h1>
