--- src/app/order-list/order-list.component.scss
@@ -1,45 +1,0 @@
-mat-card {
-    min-width: 150px;
-    max-width: 500px;
-    width: 100%;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1); /* Add shadow to all mat-cards */
-  }
-  
-  .card-container {
-    display: flex;
-    flex-wrap: wrap;
-    gap: 16px; 
-  }
-  
-  .example-form:not(:first-child) {
-    margin-top: 16px;
-  }
-  
-  .example-full-width {
-    width: 100%;
-  }
-  
-  .example-form:not(:first-child) {
-    margin-top: 16px;
-  }
-  
-  td {
-    padding-right: 8px;
-  }
-  
-  .order-table {
-    width: 100%;
-    border-collapse: collapse;
-  }
-  
-  .order-table th,
-  .order-table td {
-    padding: 8px;
-    border: 1px solid #ccc;
-  }
-  
-  .order-table th {
-    background-color: #f2f2f2;
-    text-align: left;
-  }
-
