--- src/app/orders/order-detail/order-detail.component.html
@@ -1,1 +1,0 @@
-<p>order-detail works!</p>
