+++ src/app/orders/orders.component.html
@@ -1,6 +1,4 @@
+
 <div>
-    <app-order-detail></app-order-detail>
-</div>
-<div>
-    <app-order-detail></app-order-detail>
+    <app-order-list></app-order-list>
 </div>
