--- src/app/product-list/product-list.component.html
@@ -1,77 +1,0 @@
-<!-- Add this in your component's template file -->
-<div class="card-container">
-  <mat-card>
-    <mat-card-header>
-      <h2>Products</h2>
-    </mat-card-header>
-    <mat-card-content>
-      <table class="products-table">
-        <thead>
-          <tr>
-            <th>ID</th>
-            <th>Title</th>
-            <th>Price</th>
-          </tr>
-        </thead>
-        <tbody>
-          <tr *ngFor="let product of products">
-            <td>{{ product.id }}</td>
-            <td>{{ product.title }}</td>
-            <td>{{ product.price }}</td>
-          </tr>
-        </tbody>
-      </table>
-    </mat-card-content>
-  </mat-card>
-  
-  <mat-card>
-    <mat-card-header>
-      <h2>Add New Product</h2>
-    </mat-card-header>
-    <mat-card-content>
-      <form [formGroup]="productForm" (ngSubmit)="addProduct()">
-        <div class="mat-form-field">
-          <input matInput placeholder="Title" formControlName="title">
-        </div>
-        <div class="mat-form-field">
-          <input matInput placeholder="price" formControlName="price">
-        </div>
-        <button mat-raised-button color="primary" type="submit">Add</button>
-      </form>
-    </mat-card-content> 
-  </mat-card>
-
-  <mat-card>
-    <mat-card-header>
-      <h2>Delete</h2>
-    </mat-card-header>
-    <mat-card-content>
-      <form [formGroup]="productForm" (ngSubmit)="deleteProduct()">
-        <div class="mat-form-field">
-          <input matInput type="number" placeholder="Enter ID" formControlName="id">
-        </div>
-        <button mat-raised-button color="warn" type="submit">Delete</button>
-      </form>
-    </mat-card-content>
-  </mat-card>
-
-  <mat-card>
-    <mat-card-header>
-      <h2>Edit</h2>
-    </mat-card-header>
-    <mat-card-content>
-      <form [formGroup]="productForm" (ngSubmit)="editProduct()">
-        <div class="mat-form-field">
-          <input matInput type="number" placeholder="Enter ID" formControlName="id">
-        </div>
-        <div class="mat-form-field">
-          <input matInput placeholder="Title" formControlName="title">
-        </div>
-        <div class="mat-form-field">
-          <input matInput placeholder="price" formControlName="price">
-        </div>
-        <button mat-raised-button color="accent" type="submit">Edit</button>
-      </form>
-    </mat-card-content>
-  </mat-card>
-</div>
