--- src/app/product/add-product/add-product.component.html
@@ -1,16 +1,0 @@
-<mat-card>
-    <mat-card-header>
-      <h2>Add New Product</h2>
-    </mat-card-header>
-    <mat-card-content>
-      <form [formGroup]="productForm" (ngSubmit)="addProduct()">
-        <div class="mat-form-field">
-          <input matInput placeholder="Title" formControlName="title">
-        </div>
-        <div class="mat-form-field">
-          <input matInput placeholder="Author" formControlName="author">
-        </div>
-        <button mat-raised-button color="primary" type="submit">Add</button>
-      </form>
-    </mat-card-content>
-</mat-card>
