--- src/app/products/product-detail/product-detail.component.html
@@ -1,1 +1,0 @@
-<p>product-detail works!</p>
