+++ src/app/products/products.component.html
@@ -1,8 +1,3 @@
-<div>
-    <div>
-        <app-product-detail></app-product-detail>
-    </div>
-    <div>
-        <app-order-list></app-order-list>
-    </div>
-</div>
+
+<app-product-list></app-product-list>
+
