--- src/app/top-bar/top-bar.component.html
@@ -1,3 +1,0 @@
-<a>
-    <h1>scill chain</h1>
-</a>
