--- src/app/top-bar/top-bar.component.ts
@@ -1,11 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-top-bar',
-  templateUrl: './top-bar.component.html',
-  styleUrls: ['./top-bar.component.scss']
-})
-
-export class TopBarComponent {
-  
-}
